+++ framework/db/sqlite/QueryBuilder.php
@@ -41,6 +41,7 @@ class QueryBuilder extends \yii\db\QueryBuilder
         Schema::TYPE_MONEY => 'decimal(19,4)',
     ];
 
+
     /**
      * Generates a batch INSERT SQL statement.
      * For example,
@@ -62,6 +63,12 @@ class QueryBuilder extends \yii\db\QueryBuilder
      */
     public function batchInsert($table, $columns, $rows)
     {
+        // SQLite supports batch insert natively since 3.7.11
+        // http://www.sqlite.org/releaselog/3_7_11.html
+        if (version_compare(\SQLite3::version()['versionString'], '3.7.11', '>=')) {
+            return parent::batchInsert($table, $columns, $rows);
+        }
+
         if (($tableSchema = $this->db->getTableSchema($table)) !== null) {
             $columnSchemas = $tableSchema->columns;
         } else {